requirements.txt
altair==4.2.0
appnope==0.1.3
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
importlib-metadata==5.1.0
ipdb==0.13.9
ipython==8.6.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
plotly==5.11.0
plotly-express==0.4.1
prompt-toolkit==3.0.33
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.1
statsmodels==0.13.5
streamlit==1.15.1
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wcwidth==0.2.5
zipp==3.11.0
